Move extended channel methods to trait
