+++ tests/Application/src/Entity/Channel.php
@@ -12,24 +12,10 @@ declare(strict_types=1);
 namespace Tests\BitBag\SyliusMolliePlugin\Entity;
 
 use BitBag\SyliusMolliePlugin\Entity\ChannelInterface;
-use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
+use BitBag\SyliusMolliePlugin\Entity\ChannelTrait;
 use Sylius\Component\Core\Model\Channel as BaseChannel;
-use Sylius\Component\Currency\Model\CurrencyInterface;
 
 final class Channel extends BaseChannel implements ChannelInterface
 {
-    public function getAllCurrencies(): Collection
-    {
-        $allCurrencies = new ArrayCollection([
-            $this->getBaseCurrency(),
-        ]);
-
-        /** @var CurrencyInterface $currency */
-        foreach ($this->getCurrencies() as $currency) {
-            $allCurrencies->add($currency);
-        }
-
-        return $allCurrencies;
-    }
+    use ChannelTrait;
 }
